+++ src/app/service/poap/OptInPOAP.ts
@@ -55,11 +55,20 @@ const OptInPOAP = async (user: User, dmChannel: DMChannel): Promise<void> => {
 					.setLabel('No')
 					.setStyle('SECONDARY'),
 			);
-		Log.debug('user has DMs option turned off, now asking user for opt-in to get DM POAPs');
-		const message: Message = await dmChannel.send({
-			content: 'I can send you POAPs directly to you. Would you like me to do that going forward?',
+		Log.debug('user has not opted in to DMs, now asking user for opt-in to get DM POAPs');
+		const message: Message | void = await dmChannel.send({
+			content: 'Would you like me to send you POAPs directly to you going forward?',
 			components: [row],
+		}).catch(e => {
+			LogUtils.logError('failed to ask for opt-in', e);
+			return;
 		});
+		
+		if (message == null) {
+			Log.debug('did not send opt-in message');
+			return;
+		}
+		
 		// 5 minute timeout
 		try {
 			await message.awaitMessageComponent({
@@ -96,13 +105,15 @@ const OptInPOAP = async (user: User, dmChannel: DMChannel): Promise<void> => {
 					isDMEnabled: true,
 				},
 			});
-			await message.edit({ content: 'Direct messages enabled! You will now receive POAPs, thank you!', components: [] });
+			await message.edit({ content: 'Direct messages enabled! I will send you POAPs as soon as I get them, thank you!', components: [] });
 			Log.debug('user settings updated');
 		}
 		Log.debug('user settings update skipped');
 	} else {
 		Log.debug(`user is opted in to dms, userId: ${user.id}`);
-		await dmChannel.send({ content: 'I will send you POAPs as soon as I get them!' });
+		await dmChannel.send({ content: 'I will send you POAPs as soon as I get them!' }).catch(e => {
+			LogUtils.logError('failed to send opt-in confirmation', e);
+		});
 	}
 };