+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 2.7.0-RELEASE (2022-01-18)
+
+1. Add more stability to opt-in messages
+2. Add account commands
+   - /account link (renamed from /account verify)
+   - /account list
+   - /account unlink
+3. Restrict DEGEN to certain discord servers
+4. Upgrade discord.js -> 13.6.0, upgrade slash-create -> 5.0.3
+5. Introduce basic `/claim` command and prompt user for opt-in on slash command
+6. Increase poap max time to 12 hours
+7. Add poap expiration cron job
+8. Enhance poap distribution to with ephemeral
+   - fix timeout reply after poap distribution
+   - enhance poap distribution loop
+   - enhance poap end
+9. Parse blank strings for msg embed display
+10. Prompt users to DM delivery is /claim is executed from channel
+11. Message enhancements to twitter flow
+
 ## 2.6.2-RELEASE (2022-01-13)
 
 1. Handle twitter spaces exceptions